% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocChallenges-package.R
\docType{package}
\name{BiocChallenges-package}
\alias{BiocChallenges}
\alias{BiocChallenges-package}
\title{BiocChallenges: Challenges for the Bioconductor community}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package hosts challenges contributed by and for the Bioconductor community.
    It provides functionality to manage, filter, and display challenges as articles of a pkgdown website.
    Challenges are bite-sized projects led by volunteers, encouraging collaboration and sharing of expertise between contributors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kevinrue.github.io/BiocChallenges/}
  \item \url{https://github.com/kevinrue/BiocChallenges}
  \item Report bugs at \url{https://github.com/kevinrue/BiocChallenges/issues}
}

}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevinrue67@gmail.com} (\href{https://orcid.org/0000-0003-3899-3872}{ORCID})

}
\keyword{internal}
