% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.R
\name{format_date_created}
\alias{format_date_created}
\title{Dates}
\usage{
format_date_created(params)
}
\arguments{
\item{params}{Challenge parameters as \code{list}.}
}
\value{
\code{format_date_created()} returns a character value indicating the date that a challenge was created.
}
\description{
Dates
}
\examples{
params <- list(date = "2020-11-08")
format_date_created(params)
}
