% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difficulty.R
\docType{data}
\name{difficulty_levels}
\alias{difficulty_levels}
\alias{format_difficulty}
\title{Challenge Difficulty Levels}
\format{
\code{difficulty_levels} is a character vector of valid difficulty levels.
}
\usage{
difficulty_levels

format_difficulty(params)
}
\arguments{
\item{params}{Challenge parameters as \code{list}.}
}
\value{
\code{format_difficulty()} returns a character value indicating the challenge difficulty.
}
\description{
Challenge Difficulty Levels
}
\examples{
difficulty_levels

params <- list(difficulty = "novice")
cat(format_difficulty(params))
}
\keyword{datasets}
