% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaders.R
\name{format_leaders}
\alias{format_leaders}
\title{Challenge Leaders}
\usage{
format_leaders(params)
}
\arguments{
\item{params}{Challenge parameters as \code{list}.}
}
\value{
\code{format_leaders()} returns a character value indicating the challenge leaders.
}
\description{
Challenge Leaders
}
\examples{
leaders <- list(
    kevinrue = "Kevin Rue-Albrecht"
)
params <- list(leaders = list(value = leaders))
cat(format_leaders(params), sep = "\n")
}
