% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{format_topics}
\alias{format_topics}
\title{Challenge Topics}
\usage{
format_topics(params)
}
\arguments{
\item{params}{Challenge parameters as \code{list}.}
}
\value{
\code{format_topics()} returns a character value indicating the challenge topic tags.
}
\description{
Challenge Topics
}
\examples{
topic_tags

params <- list(topics = c("Challenges", "Community"))
cat(format_topics(params))
}
