% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_repositories_info_table}
\alias{gh_repositories_info_table}
\title{Tabulate Information for GitHub Repositories}
\usage{
gh_repositories_info_table(repos)
}
\arguments{
\item{repos}{Character vector of GitHub repositories, each formatted as \code{owner/repo}.}
}
\value{
A \code{data.frame} with the following columns:
\describe{
\item{\code{repository}}{HTML content representing a link to the GitHub repository.}
\item{\code{stargazers}}{Integer count of stargazers on the GitHub repository.}
}
}
\description{
Tabulate Information for GitHub Repositories
}
\examples{
if (interactive()) {
    gh_repositories_info_table("kevinrue/BiocChallenges")
}
}
