% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_stargazer_count}
\alias{gh_stargazer_count}
\title{Query Information from GitHub API}
\usage{
gh_stargazer_count(repo)
}
\arguments{
\item{repo}{GitHub repository, formatted as \code{owner/repo}.}
}
\value{
\code{gh_stargazer_count()} returns the count of stargazers for a GitHub repository.
}
\description{
Query Information from GitHub API
}
\examples{
if (interactive()) {
    gh_stargazer_count("kevinrue/BiocChallenges")
}
}
