% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{difficulty_levels}
\alias{difficulty_levels}
\alias{format_difficulty}
\title{Challenge Difficulty}
\format{
\code{difficulty_levels} is a character vector of valid difficulty levels.


}
\usage{
difficulty_levels

format_difficulty(params)
}
\arguments{
\item{param}{Challenge parameters as \code{list}.}
}
\value{
\code{format_difficulty()} returns a character value indicating the challenge difficulty.
}
\description{
Challenge Difficulty
}
\examples{
difficulty_levels

params <- list(difficulty = "novice")
format_difficulty(params)
}
\keyword{datasets}
