% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\docType{data}
\name{topic_tags}
\alias{topic_tags}
\alias{format_topics}
\title{Challenge Topics}
\format{
\code{topic_tags} is a character vector of valid topic tags.
}
\usage{
topic_tags

format_topics(params)
}
\arguments{
\item{params}{Challenge parameters as \code{list}.}
}
\value{
\code{format_topics()} returns a character value indicating the challenge topic tags.
}
\description{
Challenge Topics
}
\examples{
topic_tags

params <- list(topics = c("Challenges", "Community"))
cat(format_topics(params))
}
\keyword{datasets}
